refactor: split agr_mati into entity/interfaces/api modules and fix Maven Central publish
Convert agr_mati from a single Quarkus module into a Maven aggregator
(org.alliancegenome:agr_mati, packaging=pom) with three child modules so
consumers can depend only on what they need:

- entity      org.alliancegenome.mati.entity  (Identifier, IdentifiersRange,
              SubdomainEntity) - Quarkus/JPA deps are 'provided', non-transitive
- interfaces  org.alliancegenome.mati.interfaces (REST interface definitions)
- api         the Quarkus app, repackaged to org.alliancegenome.mati.api.*,
              depends on entity + interfaces

The entity/interfaces package names are unchanged so existing consumers'
imports stay valid; only the api module is repackaged.

Fix the broken Maven Central publish: OSSRH (s01.oss.sonatype.org) reached
end-of-life 2025-06-30. Mirror agr_curation's working setup - drop
distributionManagement + nexus-staging-maven-plugin and publish via
org.sonatype.central:central-publishing-maven-plugin (publishingServerId=ossrh,
autoPublish), which bundles all three modules into one deployment from a
single root 'mvn deploy'. Source/javadoc/gpg plugins moved to the parent so
every module ships the artifacts Central requires.

Also: Dockerfile runner path -> api/target/api-*-runner.jar; Makefile and
release workflow made reactor-aware (versions:set -DprocessAllModules, -pl api).
