+++ api/src/main/resources/db/migration/V0002__SCRUM-1493.sql
