+++ api/src/test/java/org/alliancegenome/mati/api/controller/IdentifierResourceITCase.java
@@ -1,0 +1,108 @@
+package org.alliancegenome.mati.api.controller;
+
+import io.quarkus.test.common.WithTestResource;
+import io.quarkus.test.junit.QuarkusIntegrationTest;
+import io.restassured.http.ContentType;
+
+import org.alliancegenome.mati.api.configuration.PostgresResource;
+import org.alliancegenome.mati.entity.IdentifiersRange;
+import org.junit.jupiter.api.*;
+
+import static io.restassured.RestAssured.given;
+
+/**
+ * Tests the /api/identifier (PUT, POST) endpoints
+ * for minting identifiers
+ */
+@QuarkusIntegrationTest
+@WithTestResource(value = PostgresResource.class)
+@TestInstance(TestInstance.Lifecycle.PER_CLASS)
+@Order(2)
+class IdentifierResourceITCase {
+    private String authorization;
+
+    @BeforeAll
+    void setup() {
+        authorization = "Bearer " + CognitoHelper.fetchCognitoToken();
+    }
+
+    @Test
+    void mintIdentifierDiseaseAnnotation() {
+        checkIdentifier(mintIdentifier("disease_annotation"));
+    }
+
+    @Test
+    void mintIdentifierPerson() {
+        checkIdentifier(mintIdentifier("person"));
+    }
+
+    @Test
+    void mintIdentifierReference() {
+        checkIdentifier(mintIdentifier("reference"));
+    }
+
+    @Test
+    void mintIdentifierResource() {
+        checkIdentifier(mintIdentifier("resource"));
+    }
+
+    @Test
+    void mintIdentifierRangeDisAnnot() {
+        IdentifiersRange range = mintIdentifierRange("disease_annotation", 4);
+        checkRange(range, 4);
+    }
+
+    @Test
+    void mintIdentifierRangePerson() {
+        IdentifiersRange range = mintIdentifierRange("person", 3);
+        checkRange(range, 3);
+    }
+
+    @Test
+    void mintIdentifierRangeReference() {
+        IdentifiersRange range = mintIdentifierRange("reference", 2);
+        checkRange(range, 2);
+    }
+
+    @Test
+    void mintIdentifierRangeResource() {
+        IdentifiersRange range = mintIdentifierRange("resource", 5);
+        checkRange(range, 5);
+    }
+
+    private void checkIdentifier(String identifier) {
+        Assertions.assertTrue(identifier.startsWith("AGRKB:"));
+        Assertions.assertEquals(21, identifier.length());
+    }
+
+    private void checkRange(IdentifiersRange range, int howMany) {
+        Assertions.assertTrue(range.getFirst().getCounter() > 0L);
+        Assertions.assertEquals(range.getLast().getCounter() - range.getFirst().getCounter() + 1, howMany);
+    }
+
+    private String mintIdentifier(String subdomain) {
+        return given().
+            contentType(ContentType.JSON).
+            header("Accept", "application/json").
+            header("Authorization", authorization).
+            header("subdomain", subdomain).
+            when().
+            put("http://localhost:8081/api/identifier").
+            then().
+            statusCode(200).
+            extract().path("curie").toString();
+    }
+
+    private IdentifiersRange mintIdentifierRange(String subdomain, int howMany) {
+        return given().
+            contentType(ContentType.JSON).
+            header("Accept", "application/json").
+            header("Authorization", authorization).
+            header("subdomain", subdomain).
+            header("value", String.valueOf(howMany)).
+            when().
+            post("http://localhost:8081/api/identifier").
+            then().
+            extract().body().as(IdentifiersRange.class);
+    }
+}
